Refactor: extract `ErrorDisplay` directory
