+++ frontend/packages/erd-core/src/components/ERDRenderer/ErrorDisplay/ParseErrorDisplay.module.css
@@ -1,36 +1,4 @@
 .wrapper {
-  font-family: var(--main-font, Inter);
-  display: flex;
-  padding: 0px var(--spacing-6, 24px);
-  flex-direction: column;
-
-  height: 100%;
-
-  justify-content: center;
-  align-items: center;
-  gap: var(--spacing-10, 40px);
-
-  background-color: var(--global-background);
-}
-
-.main {
-  display: flex;
-  padding: var(--spacing-8, 32px);
-  align-items: flex-start;
-  gap: var(--spacing-6, 24px);
-
-  border-left: 1px solid var(--callout-warning-border, rgba(255, 215, 72, 0.2));
-  background: var(--callout-warning-background, rgba(255, 215, 72, 0.1));
-}
-
-.iconWrapper {
-  display: flex;
-  width: 24px;
-  padding-top: var(--spacing-1, 4px);
-  align-items: center;
-}
-
-.inner {
   display: flex;
   flex-direction: column;
   justify-content: center;
