+++ frontend/packages/erd-core/src/components/ERDRenderer/ErrorDisplay/index.ts
@@ -1,0 +1,1 @@
+export * from './ErrorDisplay'
